🚀 feat: AppBarTitleをcomponent化して完了詳細画面にも適用
